Import default settings once when app is instantiated and wrap importing code in a method called load_config_file.
Wrap the code for loading the config.json file in a method called load_config_file.  This is called once when the app is instantiated, and current instance of the app has attribute called default_settings which contains all the default values.  This attribute is referenced any time a default setting is needed.