Channels view apply button applies both channels diameter and needle length.
Now needle length is set in the config_values when user hits "apply".  This is the value that is used in the export pdf, rather than grabbing the current value from the spin box.   This means that to adjust the needle length, the user must press "apply" for their changes to work.